requirements.txt
numpy==1.24.4
pillow==10.4.0
torch==2.1.2
torchvision==0.16.2
gradio
scikit-learn